code: Add ES6 Tooling - 3. ES6 Modules
