+++ js-basics/ES6-Tooling/CommonJS Modules/after/circle.js
@@ -1,0 +1,15 @@
+// Implementation Detail 
+const _radius = new WeakMap();
+
+// Public Interface
+class Circle {
+  constructor(radius) {
+    _radius.set(this, radius);
+  }
+
+  draw() {
+    console.log('Circle with radius ' + _radius.get(this));
+  }
+}
+
+module.exports = Circle;
