+++ js-basics/ES6-Tooling/CommonJS Modules/after/index.js
@@ -1,0 +1,5 @@
+
+const Circle = require('./circle');
+
+const c = new Circle(10);
+c.draw();
