+++ js-basics/ES6-Tooling/CommonJS Modules/before/index.js
@@ -1,0 +1,15 @@
+
+const _radius = new WeakMap();
+
+class Circle {
+  constructor(radius) {
+    _radius.set(this, radius);
+  }
+
+  draw() {
+    console.log('Circle with radius ' + _radius.get(this));
+  }
+}
+
+const c = new Circle(10);
+c.draw();
